refactor: cleaned up repo
