+++ app.py
@@ -1,14 +1,15 @@
 from flask import Flask, jsonify, redirect, request
 from flask_cors import CORS
-import bill_tracker_core as core
-import db_interactions as database
-import parlpy.utils.constituency as pp_constituency
-import email_sender
+
+import insight
+import insight.parliament
+from gcp_util import database
+from parlpy.utils.constituency import get_constituencies_from_post_code
+from communications import email, notification
 import logging
 import string
 import random
 import requests
-import notifications
 
 app = Flask(__name__)
 logger = logging.getLogger()
@@ -35,7 +36,7 @@ def register():
     :return: Session token if successful, an Error otherwise.
     """
     # Get new User details from form:
-    email = request.form['email']
+    email_address = request.form['email']
     password = request.form['password']  # The given password is already hashed
     notification_token = request.form['notification_token']
     postcode = request.form['postcode']
@@ -46,18 +47,19 @@ def register():
         return jsonify({"error": "notification_token_error"})
     if not is_valid_postcode(postcode):  # Check that the postcode is valid
         return jsonify({"error": "postcode_error"})
-    if not email_sender.is_valid_email(email):  # Check that the given email is a valid email address
+    if not email.is_valid_email(email_address):  # Check that the given email is a valid email address
         return jsonify({"error": "email_error"})
 
-    if not is_new_address(email):  # Check if the given email is already in use
+    if not is_new_address(email_address):  # Check if the given email is already in use
         return jsonify({"error": "email_in_use_error"})
 
     # Add new user to the database:
-    new_user = core.User(email, password, notification_token, postcode, create_session_token())  # Create new user
+    new_user = insight.User(email_address, password, notification_token, postcode,
+                            create_session_token())  # Create new user
     add_user_to_database(new_user)  # Add new User to the database
     # Send email to user's email address
-    email_sender.send_email(new_user.email, "Insight: Registration", "Thanks for registering to use the Insight app!"
-                                                                     "\n\n-The Insight team")
+    email.send_email(new_user.email, "Insight: Registration", "Thanks for registering to use the Insight app!"
+                                                              "\n\n-The Insight team")
     # Return the session token
     return jsonify({"session_token": new_user.session_token})
 
@@ -69,18 +71,18 @@ def login():
     :return: Message with the user's unique token if successful, or an error message explaining why login failed.
     """
     # Get form information:
-    email = request.form['email']
+    email_address = request.form['email']
     password = request.form['password']
-    if is_new_address(email):
+    if is_new_address(email_address):
         return jsonify({"error": "new_email_error"})  # Email does not correspond to a User
     # Get user from database using username, check if user is valid.
-    user = fetch_user(email)  # Construct the user object
+    user = fetch_user(email_address)  # Construct the user object
     if user and user.verify_password(password):
         # Send email to user address informing of new login
-        email_sender.send_email(user.email, "Insight: new login", "A new device signed in to your Insight account. We'"
-                                                                  "re sending you this email to make sure it was you!"
-                                                                  " If it wasn't, please respond to this email to let "
-                                                                  "us know!\n\n-The Insight team")
+        email.send_email(user.email, "Insight: new login", "A new device signed in to your Insight account. We'"
+                                                           "re sending you this email to make sure it was you!"
+                                                           " If it wasn't, please respond to this email to let "
+                                                           "us know!\n\n-The Insight team")
         return jsonify({"session_token": user.session_token})  # Return the session token
     # Return the session token
     return jsonify({"error": "incorrect_password_error"})  # Given wrong password
@@ -93,9 +95,9 @@ def login_with_token():
     :return: Message indicating the login was successful, or an error message explaining why it was not successful.
     """
     # Get form information:
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
-    if verify_user(email, session_token):  # Verify the user using email and session token
+    if verify_user(email_address, session_token):  # Verify the user using email and session token
         return jsonify({"success": "login_successful"})  # Return success message
     return jsonify({"error": "login_unsuccessful"})  # Given the wrong token
 
@@ -108,11 +110,11 @@ def get_bill():
     :return: The bill, in a suitable format if successful, an error message otherwise.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     bill_id = request.form['bill_id']  # The id of the bill to fetch
 
-    if not verify_user(email, session_token):  # Verify the user
+    if not verify_user(email_address, session_token):  # Verify the user
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     bill = fetch_bill(str(bill_id))  # Fetch and construct the bill with the given id
@@ -120,11 +122,15 @@ def get_bill():
     if not bill:
         return jsonify({"error": "query_failed"})  # Query failed, no such bill exists
 
-    bill_dict = bill.to_dict()  # Convert the bill to a suitable format to be transmitted
-    user_interactions = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
-    bill_dict['likes'] = user_interactions[0]  # Get number of likes
-    bill_dict['dislikes'] = user_interactions[1]  # Get number of dislikes
-    bill_dict['user_vote'] = fetch_user_interaction(fetch_user_id(email), bill.id)
+    likes, dislikes = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
+    bill_dict = prepare_bill(
+        bill,
+        {
+            "likes": likes,
+            "dislikes": dislikes,
+            "user_vote": fetch_user_interaction(fetch_user_id(email_address), bill.id),
+        }
+    )
 
     return jsonify(bill_dict)  # Return the Bill as a dictionary
 
@@ -137,10 +143,10 @@ def get_bills():
     :return: The bills in a suitable format if successful, otherwise an error message.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
 
-    if not verify_user(email, session_token):  # Verify the user
+    if not verify_user(email_address, session_token):  # Verify the user
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     bill_id_list = fetch_recent_bills()  # Get the 50 most recent bills
@@ -152,11 +158,15 @@ def get_bills():
     for bill_id in bill_id_list:
         bill = fetch_bill(str(bill_id[0]))  # Fetch and construct the bill with the given id
         if bill:
-            bill_dict = bill.to_dict()  # Convert the bill to a suitable format to be transmitted
-            user_interactions = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
-            bill_dict['likes'] = user_interactions[0]  # Get number of likes
-            bill_dict['dislikes'] = user_interactions[1]  # Get number of dislikes
-            bill_dict['user_vote'] = fetch_user_interaction(fetch_user_id(email), bill.id)
+            likes, dislikes = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
+            bill_dict = prepare_bill(
+                bill,
+                {
+                    "likes": likes,
+                    "dislikes": dislikes,
+                    "user_vote": fetch_user_interaction(fetch_user_id(email_address), bill.id),
+                }
+            )
             bill_list.append(bill_dict)  # Add the bill to the bill list
 
     if not bill_list:
@@ -173,11 +183,11 @@ def get_mp_bills():
     :return: A list of bills voted by the given MP, if successful, or an error message if it failed.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     mp_id = request.form['mp_id']  # Get ParliamentMember id
 
-    if not verify_user(email, session_token):  # Verify the user
+    if not verify_user(email_address, session_token):  # Verify the user
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     bill_query = database.select(f"SELECT DISTINCT Bills.billID, titleStripped, shortDesc, dateAdded, expiration, "
@@ -189,14 +199,20 @@ def get_mp_bills():
 
     bill_list = []  # Holds the list of bills to be transmitted
     for bill_data in bill_query:  # Put all bills in the query in the correct format
-        bill = core.Bill(bill_data[0], bill_data[1], None, str(bill_data[3])[:10].replace(" ", ""),
-                         bill_data[4], bill_data[6], parse_text(bill_data[2]), link=bill_data[5])  # Create MP object
-        bill_dict = bill.to_dict()  # Get the dictionary representation of the bill
-        user_interactions = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
-        bill_dict['likes'] = user_interactions[0]  # Get number of likes
-        bill_dict['dislikes'] = user_interactions[1]  # Get number of dislikes
-        bill_dict['user_vote'] = fetch_user_interaction(fetch_user_id(email), bill.id)
-        bill_list.append(bill_dict)  # Append the bill to the list
+        bill = insight.parliament.Bill(bill_data[0], bill_data[1], None, str(bill_data[3])[:10].replace(" ", ""),
+                                       bill_data[4], bill_data[6], strip_text(bill_data[2]),
+                                       link=bill_data[5])  # Create MP object
+
+        likes, dislikes = fetch_user_interactions(bill.id)  # Get the user interactions for the bill
+        bill_dict = prepare_bill(
+            bill,
+            {
+                "likes": likes,
+                "dislikes": dislikes,
+                "user_vote": fetch_user_interaction(fetch_user_id(email_address), bill.id),
+            }
+        )
+        bill_list.append(bill_dict)  # Add the bill to the bill list
     return jsonify(bill_list)  # Return the list of bills
 
 
@@ -208,21 +224,22 @@ def send_message():
     :return: A success message, if the email was sent successfully, otherwise an error message
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     # Get information to send email
     mp_id = request.form['mp_id']
     message = request.form['message']
     # Verify the user:
-    if not verify_user(email, session_token):  # Verify the user
+    if not verify_user(email_address, session_token):  # Verify the user
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     mp = fetch_mp(mp_id)  # Construct and return the parliament member by following given id
 
     if mp:  # If the MP was successfully constructed
         try:
-            email_sender.send_email(mp.email, "Insight Update!", message)  # Send the email
+            email.send_email(mp.email, "Insight Update!", message)  # Send the email
             return jsonify({"success": "email_sent"})  # If sent without errors, return success message
+        # todo: can this be more specific?
         except Exception:
             return jsonify({"error": "email_failed_to_send"})  # Error with mail sending
 
@@ -236,18 +253,19 @@ def update_postcode():
     :return: Success statement if the postcode was changed, an error statement otherwise.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     postcode = request.form['postcode']  # Get the new postcode
 
     # Verify the user:
-    if not verify_user(email, session_token):
+    if not verify_user(email_address, session_token):
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     if not is_valid_postcode(postcode):  # Check that the postcode is valid
         return jsonify({"error": "postcode_error"})
     try:
-        database.interact(f"UPDATE Users SET postcode='{postcode}' WHERE email='{email}';")  # Update user's postcode
+        database.interact(
+            f"UPDATE Users SET postcode='{postcode}' WHERE email='{email_address}';")  # Update user's postcode
     except RuntimeWarning:
         return jsonify({"error": "query_error"})  # Error when executing sql statement
 
@@ -261,13 +279,13 @@ def get_mp_votes():
     :return: A list of the MP's votes.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     # Get information to send email
     mp_id = request.form['mp_id']
 
     # Verify the user:
-    if not verify_user(email, session_token):
+    if not verify_user(email_address, session_token):
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
     if not fetch_mp(mp_id):  # Check if mp_id exists
@@ -297,13 +315,13 @@ def get_local_mp():
     :return: The user's local MP if successful, an error message otherwise.
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     # Verify the user:
-    if not verify_user(email, session_token):
+    if not verify_user(email_address, session_token):
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
-    user = fetch_user(email)  # Get the user's details
+    user = fetch_user(email_address)  # Get the user's details
     if not user:
         return jsonify({"error": "user_error"})  # Verification unsuccessful
 
@@ -328,15 +346,15 @@ def set_user_vote():
     :return: A success message, if no errors occurred, otherwise an error message
     """
     # Get user info for verification
-    email = request.form['email']
+    email_address = request.form['email']
     session_token = request.form['session_token']
     bill_id = request.form['bill_id']
     positive = request.form['positive']
     # Verify the user:
-    if not verify_user(email, session_token):
+    if not verify_user(email_address, session_token):
         return jsonify({"error": "invalid_credentials"})  # Verification unsuccessful
 
-    user_id = fetch_user_id(email)  # Fetch and construct the User object from the database
+    user_id = fetch_user_id(email_address)  # Fetch and construct the User object from the database
     vote_state = fetch_user_interaction(user_id, bill_id)  # Gets the current reaction state of the bill for the user
 
     # Construct the appropriate SQL statement
@@ -359,6 +377,22 @@ def set_user_vote():
 
 # ////// End region ////// Helper functions //////
 
+def prepare_bill(bill: insight.parliament.Bill, additional_values: {} = None) -> dict:
+    """
+    Prepares bill to be sent to front end, adding any extra fields that are expected (such as likes)
+    :param bill: Bill to prepare
+    :param additional_values: Extra key-value pairs to add to result
+    :return: Bill in dict form, with extra fields.
+    """
+    bill_dict = bill.to_dict()
+
+    # Add extra values, if any are given
+    if additional_values is not None:
+        for key in additional_values:
+            bill_dict[key] = additional_values[key]
+
+    return bill_dict
+
 
 def create_session_token() -> str:
     """
@@ -373,16 +407,16 @@ def create_session_token() -> str:
     return token  # Return the unique token
 
 
-def verify_user(email: str, session_token: str) -> bool:
+def verify_user(email_address: str, session_token: str) -> bool:
     """
     Verify the user using their email and token. Checks if the email address is used, and verifies the token.
-    :param email: The email address of the user.
+    :param email_address: The email address of the user.
     :param session_token: The session token of the user.
     :return: True if the user was verified, False otherwise.
     """
-    if is_new_address(email):  # Check if the email corresponds to a User
+    if is_new_address(email_address):  # Check if the email corresponds to a User
         return False
-    user = fetch_user(email)  # Get the user form the database using their email
+    user = fetch_user(email_address)  # Get the user form the database using their email
     if user and user.verify_token(session_token):
         return True  # Login successful
     return False  # Tokens do not match
@@ -400,45 +434,19 @@ def is_new_address(email_address: str) -> bool:
     return True  # If the query returns an empty list return True
 
 
-def is_valid_postcode(postcode: str) -> bool:
-    """
-    Checks if the given postcode is a valid postcode using the https://api.postcodes.io/ API.
-    :param postcode: The postcode to evaluate.
-    :return: True if the postcode is valid, False otherwise.
-    """
-    url = f"https://api.postcodes.io/postcodes/{postcode}/validate"  # Use the postcode.io API to validate postcodes
-    response = requests.get(url).json()
-    if response['status'] == 200:  # Request was processed properly
-        return response['result']  # Get the API's evaluation
-    return False  # API refused to evaluate the request (typeError)
-
-
-def add_user_to_database(user: core.User) -> None:
+def add_user_to_database(user: insight.User) -> None:
     """
     Add the given User object to the database.
     :param user: User object.
     :return: None.
     """
-    if not user or type(user) is not core.User:  # If the given user is not a User object, return immediately
+    if not user or type(user) is not insight.User:  # If the given user is not a User object, return immediately
         return
     # The SQL statement to add the user into the Users table:
     statement = f"INSERT INTO Users (email,password,postcode,sessionToken,notificationToken) VALUES ('{user.email}','" \
                 f"{user.password_hash}','{user.postcode}','{user.session_token}','{user.notification_token}');"
     database.interact(statement)  # Carry out the SQL statement
     return
-
-
-def parse_text(text: str) -> str:
-    """
-    Finds and removes the escape characters in the given string. Checks for linux and windows escape characters.
-    :param text: The string to be parsed.
-    :return: The parsed string.
-    """
-    if "\r" in text:
-        text = text.replace("\r", "")  # Remove linux next line char
-    if "\n" in text:
-        text = text.replace("\n", "")  # Remove mac & windows next line char
-    return text
 
 
 def notify_users(title: str, body: str) -> None:
@@ -457,11 +465,11 @@ def notify_users(title: str, body: str) -> None:
         user = fetch_user(user_email)  # Construct the User object
         if user:
             user_list.append(user)  # Add User to the list of users
-    notifications.send_notification_to_clients(user_list, title, body)  # Build and send the notification to all users
+    notification.send_notification_to_clients(user_list, title, body)  # Build and send the notification to all users
     return
 
 
-def fetch_user(email_address: str) -> core.User or None:
+def fetch_user(email_address: str) -> insight.User or None:
     """
     Finds the user with the given email address, constructs and returns the User object.
     :param email_address: The email address of the user.
@@ -471,11 +479,11 @@ def fetch_user(email_address: str) -> core.User or None:
     user = None
     if user_query:
         user_info = user_query[0]  # Get the user information
-        user = core.User(user_info[1], user_info[2], user_info[4], user_info[3], user_info[5])  # Construct user
+        user = insight.User(user_info[1], user_info[2], user_info[4], user_info[3], user_info[5])  # Construct user
     return user
 
 
-def fetch_mp(mp_id: int) -> core.ParliamentMember or None:
+def fetch_mp(mp_id: int) -> insight.parliament.Member or None:
     """
     Finds the member of parliament with the given ID, constructs and returns the ParliamentMember object.
     :param mp_id: The id of the MP.
@@ -486,12 +494,12 @@ def fetch_mp(mp_id: int) -> core.ParliamentMember or None:
     if mp_query:
         mp_info = mp_query[0]  # Extract the MP information
         # Construct MP object:
-        parliament_member = core.ParliamentMember(mp_info[0], mp_info[1], mp_info[2], mp_info[3], mp_info[4],
-                                                  mp_info[5], mp_info[6], mp_info[7], mp_info[8], mp_info[9])
+        parliament_member = insight.parliament.Member(mp_info[0], mp_info[1], mp_info[2], mp_info[3], mp_info[4],
+                                                      mp_info[5], mp_info[6], mp_info[7], mp_info[8], mp_info[9])
     return parliament_member
 
 
-def fetch_bill(bill_id: str) -> core.Bill or None:
+def fetch_bill(bill_id: str) -> insight.parliament.Bill or None:
     """
     Finds the bill with the given ID, constructs and returns the Bill object.
     :param bill_id: The id of the bill to fetch.
@@ -502,8 +510,10 @@ def fetch_bill(bill_id: str) -> core.Bill or None:
     bill = None
     if bill_query:  # If the query was successful (the bill exists), build the Bill object
         bill_data = bill_query[0]  # Get the bill's information
-        bill = core.Bill(bill_data[0], bill_data[1], bill_data[7], str(bill_data[3])[:10].replace(" ", ""),
-                         bill_data[4], bill_data[6], parse_text(bill_data[2]), bill_data[5])  # Construct the Bill
+        bill = insight.parliament.Bill(bill_data[0], bill_data[1], bill_data[7],
+                                       str(bill_data[3])[:10].replace(" ", ""),
+                                       bill_data[4], bill_data[6], strip_text(bill_data[2]),
+                                       bill_data[5])  # Construct the Bill
 
     return bill  # Return the Bill object
 
@@ -534,7 +544,7 @@ def fetch_user_id(email_address: str) -> str:
     return user_query[0][0]  # Return the user's id
 
 
-def fetch_user_interactions(bill_id: str) -> tuple:
+def fetch_user_interactions(bill_id: str) -> (int, int):
     """
     Fetches the users' votes (interaction) on the given bill.
     :param bill_id: The id of the bill.
@@ -585,7 +595,7 @@ def fetch_mp_by_postcode(postcode: str) -> int:
     :param postcode: The postcode being searched.
     :return: The id of the MP for the constituency.
     """
-    constituency = pp_constituency.get_constituencies_from_post_code(postcode)  # Get the constituency from ParlPy
+    constituency = get_constituencies_from_post_code(postcode)  # Get the constituency from ParlPy
     if not constituency:
         raise KeyError(f"Found no constituencies for postcode '{postcode}'.")  # No constituency exists, raise an error
     return constituency[0]["currentRepresentation"]["member"]["value"]["id"]  # Return the MP for the constituency
@@ -608,6 +618,38 @@ def clean_mp_votes(bill_votes: list) -> list:
         clean_votes.append(bill)  # Add the bill to the cleaned list
         prev_id = bill[0]  # Update the previous id for next iteration
     return clean_votes
+
+
+def is_valid_postcode(postcode: str) -> bool:
+    """
+    Checks if the given postcode is a valid postcode using the https://api.postcodes.io/ API.
+    :param postcode: The postcode to evaluate.
+    :return: True if the postcode is valid, False otherwise.
+    """
+    # Use the postcode.io API to validate postcodes
+    url = f"https://api.postcodes.io/postcodes/{postcode}/validate"
+    response = requests.get(url).json()
+
+    # Request was processed properly
+    if response['status'] == 200:
+        # Get the API's evaluation
+        return response['result']
+
+    # API refused to evaluate the request (typeError)
+    return False
+
+
+def strip_text(text: str) -> str:
+    """
+    Finds and removes the escape characters in the given string. Checks for linux and windows escape characters.
+    :param text: The string to be parsed.
+    :return: The parsed string.
+    """
+    if "\r" in text:
+        text = text.replace("\r", "")  # Remove linux next line char
+    if "\n" in text:
+        text = text.replace("\n", "")  # Remove mac & windows next line char
+    return text
 
 
 if __name__ == '__main__':
